+++ shared/i18n/locales/en_US/translation.json
@@ -238,9 +238,6 @@
   "Sorry, an unrecoverable error occurred{{notified}}. Please try reloading the page, it may have been a temporary glitch.": "Sorry, an unrecoverable error occurred{{notified}}. Please try reloading the page, it may have been a temporary glitch.",
   "our engineers have been notified": "our engineers have been notified",
   "Show detail": "Show detail",
-  "Revision deleted": "Revision deleted",
-  "Current version": "Current version",
-  "{{userName}} edited": "{{userName}} edited",
   "{{userName}} archived": "{{userName}} archived",
   "{{userName}} restored": "{{userName}} restored",
   "{{userName}} deleted": "{{userName}} deleted",
@@ -316,6 +313,9 @@
   "Reaction picker": "Reaction picker",
   "Could not load reactions": "Could not load reactions",
   "Reaction": "Reaction",
+  "Revision deleted": "Revision deleted",
+  "Current version": "Current version",
+  "{{userName}} edited": "{{userName}} edited",
   "Results": "Results",
   "No results for {{query}}": "No results for {{query}}",
   "Manage": "Manage",
